api/requirements.txt
aiohttp==3.0.1
aioredis==1.0.0
amqp==2.2.2
asgiref==2.1.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
channels==2.0.2
channels-redis==2.1.0
chardet==3.0.4
constantly==15.1.0
coverage==4.4.2
daphne==2.0.4
dj-database-url==0.4.2
Django==2.0.3
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-cors-headers==2.0.2
django-debug-toolbar==1.9.1
django-extensions==1.9.8
django-jet==1.0.6
django-jsonview==1.1.0
django-redis==4.9.0
django-rest-framework==0.1.0
django-watchman==0.15.0
djangorestframework==3.8.2
djangorestframework-camel-case==0.2.0
djangorestframework_simplejwt==3.2.3
fake-useragent==0.1.10
flower==0.9.2
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
idna-ssl==1.0.1
incremental==17.5.0
kombu==4.1.0
msgpack==0.5.6
multidict==4.1.0
oauthlib==2.0.6
psycopg2==2.7.3.2
PyJWT==1.5.3
python-magic==0.4.15
pytz==2018.3
redis==2.10.6
requests==2.18.4
selenium==3.9.0
six==1.11.0
sqlparse==0.2.4
Twisted==17.9.0
txaio==2.9.0
typing==3.6.2
urllib3==1.22
vine==1.1.4
yarl==1.1.1
zope.interface==4.4.3
